+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+- [Added restore provider after page reload](https://github.com/multiversx/mx-sdk-dapp-core/pull/19)
 - [Added signMessage](https://github.com/multiversx/mx-sdk-dapp-core/pull/18)
 
 ## [[0.0.0-alpha.9]](https://github.com/multiversx/mx-sdk-dapp-core)] - 2024-08-29